Lobby Scene / Boss Scene work
Boss Scene:
Added a SceneManager to allow integration into the game.
Added a tool in ObjectiveScript that would allow a scene to end in situations not involving kill counts or timers by manually ending it.

Lobby Scene:
Added the lobby scene.
Added a script that would help procedurally generate the background.

Animation work:
Added crouching. For now, it's just visual.